fixed builds outside source directory not finding GLSL shaders fixed further 64bit warnings
git-svn-id: https://camplinux.in.tum.de/svn/campvis/trunk@479 bb408c1c-ae56-11e1-83d9-df6b3e0c105e

+++ core/properties/optionproperty.h
@@ -209,7 +209,7 @@ namespace campvis {
     void campvis::GenericOptionProperty<T>::selectById(const std::string& id) {
         for (size_t i = 0; i < _options.size(); ++i) {
             if (_options[i]._id == id) {
-                setValue(i);
+                setValue(static_cast<int>(i));
                 return;
             }
         }
